Technical : Add splash screen + core:navigation
