+++ composeApp/src/androidMain/kotlin/com/despaircorp/trackshift/MainActivity.kt
@@ -4,9 +4,8 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
 import com.despaircorp.design_system.theme.TrackShiftTheme
+import com.despaircorp.trackshift.main.TrackShiftApp
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -15,16 +14,8 @@ class MainActivity : ComponentActivity() {
 
         setContent {
             TrackShiftTheme {
-
+                TrackShiftApp()
             }
         }
-    }
-}
-
-@Preview
-@Composable
-fun AppAndroidPreview() {
-    TrackShiftTheme {
-
     }
 }
