+++ core/design-system/src/commonMain/kotlin/com/despaircorp/design_system/theme/Type.kt
@@ -1,6 +1,7 @@
 package com.despaircorp.design_system.theme
 
 
+import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
