+++ core/navigation/.gitignore
