+++ core/navigation/src/androidHostTest/kotlin/com/despaircorp/navigation/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package com.despaircorp.navigation
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
