+++ core/navigation/src/androidMain/AndroidManifest.xml
