+++ core/navigation/src/androidMain/kotlin/com/despaircorp/navigation/Platform.android.kt
@@ -1,0 +1,3 @@
+package com.despaircorp.navigation
+
+actual fun platform() = "Android"
