+++ core/navigation/src/commonMain/kotlin/com/despaircorp/navigation/NavigationRoute.kt
@@ -1,0 +1,5 @@
+package com.despaircorp.navigation
+
+sealed class NavigationRoute(val route: String) {
+    object Splash : NavigationRoute("splash")
+}
