+++ core/navigation/src/commonMain/kotlin/com/despaircorp/navigation/Platform.kt
@@ -1,0 +1,3 @@
+package com.despaircorp.navigation
+
+expect fun platform(): String
