+++ core/navigation/src/iosMain/kotlin/com/despaircorp/navigation/Platform.ios.kt
@@ -1,0 +1,3 @@
+package com.despaircorp.navigation
+
+actual fun platform() = "iOS"
