--- core/ui/src/androidMain/kotlin/com/despaircorp/ui/Platform.android.kt
@@ -1,3 +1,0 @@
-package com.despaircorp.ui
-
-actual fun platform() = "Android"
