--- core/ui/src/commonMain/kotlin/com/despaircorp/ui/Platform.kt
@@ -1,3 +1,0 @@
-package com.despaircorp.ui
-
-expect fun platform(): String
