--- core/ui/src/iosMain/kotlin/com/despaircorp/ui/Platform.ios.kt
@@ -1,3 +1,0 @@
-package com.despaircorp.ui
-
-actual fun platform() = "iOS"
