+++ feature-splash-screen/src/commonMain/kotlin/com/despaircorp/feature_splash_screen/ui/SplashScreen.kt
@@ -1,0 +1,11 @@
+package com.despaircorp.feature_splash_screen.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun SplashScreen(
+    modifier: Modifier = Modifier
+) {
+
+}
